data_service_UC8_UC11/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
httpx==0.25.0
google-cloud-bigquery==3.20.1
google-auth==2.29.0
db-dtypes==1.2.0
